+++ korgw/src/jvmMain/kotlin/com/soywiz/korgw/jogl/JoglGameWindow.kt
@@ -694,7 +694,7 @@ class JvmKmlGl(val gl: GL2) : KmlGl() {
     override fun vertexAttrib3fv(index: Int, v: FBuffer): Unit = gl.glVertexAttrib3fv(index, v.nioFloatBuffer)
     override fun vertexAttrib4f(index: Int, x: Float, y: Float, z: Float, w: Float): Unit = gl.glVertexAttrib4f(index, x, y, z, w)
     override fun vertexAttrib4fv(index: Int, v: FBuffer): Unit = gl.glVertexAttrib4fv(index, v.nioFloatBuffer)
-    override fun vertexAttribPointer(index: Int, size: Int, type: Int, normalized: Boolean, stride: Int, pointer: Int): Unit = gl.glVertexAttribPointer(index, size, type, normalized, stride, pointer.toLong())
+    override fun vertexAttribPointer(index: Int, size: Int, type: Int, normalized: Boolean, stride: Int, pointer: Long): Unit = gl.glVertexAttribPointer(index, size, type, normalized, stride, pointer)
     override fun viewport(x: Int, y: Int, width: Int, height: Int): Unit = gl.glViewport(x, y, width, height)
 }
 */
